+++ core/src/main/java/com/github/klboke/kkrepo/core/RepositoryProtocol.java
@@ -1,0 +1,7 @@
+package com.github.klboke.kkrepo.core;
+
+public interface RepositoryProtocol {
+  RepositoryFormat format();
+
+  ProtocolCapability capability();
+}