+++ core/src/main/java/com/github/klboke/nexusplus/core/RepositoryProtocol.java
@@ -1,7 +1,0 @@
-package com.github.klboke.nexusplus.core;
-
-public interface RepositoryProtocol {
-  RepositoryFormat format();
-
-  ProtocolCapability capability();
-}