+++ protocol-pypi/src/main/java/com/github/klboke/kkrepo/protocol/pypi/PypiRepositoryProtocol.java
@@ -1,0 +1,17 @@
+package com.github.klboke.kkrepo.protocol.pypi;
+
+import com.github.klboke.kkrepo.core.ProtocolCapability;
+import com.github.klboke.kkrepo.core.RepositoryFormat;
+import com.github.klboke.kkrepo.core.RepositoryProtocol;
+
+public final class PypiRepositoryProtocol implements RepositoryProtocol {
+  @Override
+  public RepositoryFormat format() {
+    return RepositoryFormat.PYPI;
+  }
+
+  @Override
+  public ProtocolCapability capability() {
+    return new ProtocolCapability(true, true, true, true, true);
+  }
+}