+++ server/src/main/java/com/github/klboke/kkrepo/server/cache/NexusCacheType.java
@@ -1,0 +1,6 @@
+package com.github.klboke.kkrepo.server.cache;
+
+public enum NexusCacheType {
+  CONTENT,
+  METADATA
+}