+++ server/src/main/java/com/github/klboke/nexusplus/server/cache/NexusCacheType.java
@@ -1,6 +1,0 @@
-package com.github.klboke.nexusplus.server.cache;
-
-public enum NexusCacheType {
-  CONTENT,
-  METADATA
-}