+++ server/src/main/java/com/github/klboke/nexusplus/server/repositories/RepositoryValidationException.java
@@ -1,7 +1,0 @@
-package com.github.klboke.nexusplus.server.repositories;
-
-public class RepositoryValidationException extends RuntimeException {
-  public RepositoryValidationException(String message) {
-    super(message);
-  }
-}